+++ src/main/java/com/erayerdin/primitivefxmvc/applications/exception/OperationSystemNotSupportedException.java
@@ -1,0 +1,7 @@
+package com.erayerdin.primitivefxmvc.applications.exception;
+
+public class OperationSystemNotSupportedException extends Exception {
+    public OperationSystemNotSupportedException(String message) {
+        super(message);
+    }
+}